Move user pages and list to blueprint module
