+++ modules/users/routes.py
@@ -1,0 +1,118 @@
+from flask import Blueprint, render_template, redirect, flash, url_for, request, g
+from decorators import *
+from constants import *
+from sqlalchemy import text
+from time import strftime
+
+import auth
+import helpers as users_helpers
+
+blueprint = Blueprint('users', __name__, template_folder='templates')
+
+@blueprint.route('/')
+@login_required()
+def show_users():
+  """ Procedure to show a list of all users, with all membership details. """
+  # store which columns we want, and their displaynames
+  cols = ["user_id", "lname", "fname", "email", "matriculate_year", \
+          "grad_year", "major", "membership_desc"]
+  display = [None, "Last", "First", "Email", "Matr.", "Grad.", "Major", "Type"]
+  fieldMap = dict(zip(cols, display))
+
+  # check which table to read from
+  if 'filterType' in request.args and request.args['filterType'] == 'current':
+    tableName = 'members_current'
+    filterType = 'current'
+  elif 'filterType' in request.args and request.args['filterType'] == 'alumni':
+    tableName = 'members_alumni'
+    filterType = 'alumni'
+  else:
+    tableName = 'members'
+    filterType = 'all'
+
+  # perform query
+  query = text("SELECT * FROM " + tableName + " NATURAL JOIN membership_types")
+  results = g.db.execute(query)
+
+  # put results in a dictionary
+  result_cols = results.keys()
+  res = []
+  for result in results:
+    temp_dict = {}
+    for i, key in enumerate(result_cols):
+      if key in cols and result[i]:
+        temp_dict[key] = result[i]
+    res.append(temp_dict)
+
+  # we also want to map ids to usernames so we can link to individual pages
+  query = text("SELECT user_id, username FROM users")
+  results = g.db.execute(query)
+  idMap = dict(list(results)) # key is id, value is username
+
+  return render_template('userlist.html', data = res, fields = cols, \
+      displays=display, idMap=idMap, fieldMap=fieldMap, filterType=filterType)
+
+@blueprint.route('/view/<username>')
+@login_required()
+def show_user_profile(username):
+  """ Procedure to show a user's profile and membership details. """
+  d_dict_user, q_dict_user = users_helpers.get_user_info(username)
+  offices = users_helpers.get_office_info(username)
+  editable = users_helpers.can_edit(username)
+
+  if d_dict_user != None and q_dict_user != None:
+    return render_template('view_user.html', display = d_dict_user, \
+        info = q_dict_user, offices = list(offices), strftime = strftime,
+        perm = editable)
+  else:
+    flash("User does not exist!")
+    return redirect(url_for('home'))
+
+@blueprint.route('/edit/<username>', methods=['GET', 'POST'])
+@login_required(Permissions.UserAdmin)
+def change_user_settings(username):
+  """ Procedure to process the login page. Also handles authentication. """
+  params = {}
+  tags = ['nickname', 'usenickname', 'bday', 'email', 'email2', 'msc', 'phone', \
+      'building', 'room_num', 'major', 'isabroad']
+  tag_names = ["Nickname", "Use Nickname", "Birthday", "Email Address", \
+      "Alt. Email Address", "MSC", "Phone Number", "Building Name", "Room Number", \
+      "Major", "Is Abroad"]
+
+  # Get stored values from database
+  query = text("SELECT * FROM users NATURAL JOIN members WHERE username=:u")
+  result = g.db.execute(query, u=str(username))
+  if result.returns_rows and result.rowcount != 0:
+    result_cols = result.keys()
+    r = result.first()
+    stored_params = dict(zip(result_cols, r)) #stored_params maps sql columns to values
+
+  # Update if needed
+  if request.method == 'POST':
+
+    for (i, tag) in enumerate(tags):
+      params[tag] = request.form[tag]
+      if params[tag] and tag in ['usenickname', 'msc', 'room_num', 'isabroad']:
+        try:
+          params[tag] = int(params[tag])
+        except:
+          flash("Invalid input for field %s - your change was not saved!" % tag_names[i])
+          params[tag] = stored_params[tag]
+
+    for (i, tag) in enumerate(tags):
+      if str(params[tag]) != str(stored_params[tag]):
+
+        new_val = str(params[tag])
+
+        query = text("UPDATE members SET %s = :val WHERE user_id = :u" % tag)
+        results = g.db.execute(query, \
+            u=auth.get_user_id(username), val=new_val)
+
+        flash("%s was updated!" % tag_names[i])
+
+
+  if not params:
+    params = stored_params
+
+  return render_template('edit_user.html', params = params)
+
