+++ templates/government.html
@@ -15,7 +15,7 @@
       </tr></thead>
       {# Now, display each row #}
       {% for position in positions %}
-        <tr class="link" onclick="location.href='{{ url_for('show_user_profile',
+        <tr class="link" onclick="location.href='{{ url_for('users.show_user_profile',
           username=position['username']) }}'">
           <td>{{ position['office_name'] }}</td>
           <td>{{ position['name'] }}</td>
